+++ src/Models/Tag.php
@@ -8,6 +8,7 @@ use Hirtz\Location\Models\Collections\TagCollection;
 use Hirtz\Location\Models\Queries\LocationQuery;
 use Hirtz\Location\Models\Queries\TagQuery;
 use Hirtz\Location\Modules\ModuleTrait;
+use Hirtz\Skeleton\Models\Breadcrumb;
 use Hirtz\Skeleton\Behaviors\BlameableBehavior;
 use Hirtz\Skeleton\Behaviors\TimestampBehavior;
 use Hirtz\Skeleton\Behaviors\TrailBehavior;
@@ -220,6 +221,11 @@ class Tag extends ActiveRecord implements
     public function getAdminRoute(): array|false
     {
         return $this->id ? ['/admin/location/tag/update', 'id' => $this->id] : false;
+    }
+
+    public function getAdminIndexBreadcrumb(): Breadcrumb
+    {
+        return new Breadcrumb(Yii::t('location', 'COMMON_TAGS'), ['/admin/location/tag/']);
     }
 
     public function getPermissionName(): string